Move the client's server code to its own package and fixed run file
