+++ src/main/edu/colostate/cs/cs414/method_men/jungle/client/gui/LoginPage.java
@@ -1,5 +1,7 @@
 package edu.colostate.cs.cs414.method_men.jungle.client.gui;
 
+import edu.colostate.cs.cs414.method_men.jungle.client.socket.Login;
+
 import javax.swing.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
