+++ src/main/edu/colostate/cs/cs414/method_men/jungle/client/socket/Client.java
@@ -1,4 +1,6 @@
-package edu.colostate.cs.cs414.method_men.jungle.client.gui;
+package edu.colostate.cs.cs414.method_men.jungle.client.socket;
+
+import edu.colostate.cs.cs414.method_men.jungle.client.gui.GUI;
 
 import java.net.InetAddress;
 import java.net.Socket;
